+++ lib/packetgen/header/http/headers.rb
@@ -12,6 +12,8 @@ module PacketGen
       # @abstract Base class for HTTP headers.
       # @author Kent 'picat' Gruber
       class Headers
+        include Inspectable
+
         # Underlying Headers data (or nil).
         # @return [Hash, nil]
         attr_reader :data
@@ -56,6 +58,7 @@ module PacketGen
         def to_human
           @data
         end
+        alias format_inspect to_human
 
         # Read human-readable data to populate header data.
         # @param [String, Hash] data