+++ lib/packetgen/header/ip/addr.rb
@@ -11,6 +11,8 @@ module PacketGen
       # IP address, as a group of 4 bytes
       # @author Sylvain Daubert
       class Addr < Types::Fields
+        include Inspectable
+
         # @!attribute a1
         #  @return [Integer] IP address first byte
         define_field :a1, Types::Int8
@@ -47,6 +49,7 @@ module PacketGen
         def to_human
           fields.map { |f| self[f].to_i.to_s }.join('.')
         end
+        alias format_inspect to_human
 
         # Addr as an integer
         # @return [Integer]